+++ sources/RI.Abstractions.Common/Logging/CallbackLoggerDelegate.cs
@@ -15,5 +15,10 @@ namespace RI.Abstractions.Logging
     /// <param name="exception"> Optional exception associated with the log message. </param>
     /// <param name="format"> The log message (with optional string expansion arguments such as <c> {0} </c>, <c> {1} </c>, etc. which are expanded using <paramref name="args" />). </param>
     /// <param name="args"> Optional message arguments expanded into <paramref name="format" />. </param>
+    /// <remarks>
+    /// <note type="implement">
+    /// <paramref name="source"/>, <paramref name="exception"/>, <paramref name="format"/>, and <paramref name="args"/> can be null in any combination.
+    /// </note>
+    /// </remarks>
     public delegate void CallbackLoggerDelegate (DateTime timestampUtc, int threadId, LogLevel level, string source, Exception exception, string format, object[] args);
 }