+++ sources/RI.Abstractions.Tests/Fakes/FakeServiceD.cs
@@ -1,0 +1,9 @@
+﻿namespace RI.Abstractions.Tests.Fakes
+{
+    public sealed class FakeServiceD : IFakeServiceD
+    {
+        public FakeServiceD (IFakeServiceA serviceA, IFakeServiceB serviceB) { }
+
+        public FakeServiceD (IFakeServiceB serviceB, IFakeServiceA serviceA) { }
+    }
+}