+++ sources/RI.Abstractions.Tests/Fakes/IFakeServiceD.cs
@@ -1,0 +1,7 @@
+﻿namespace RI.Abstractions.Tests.Fakes
+{
+    public interface IFakeServiceD
+    {
+        
+    }
+}